+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Added Cargo Airship
 * You can now kick out entities using shift-right click
+* Added fall damage when crashing the aircraft (never lethal by default, configurable)
 
 # 0.4.2