+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.5.0
+
+* Added Cargo Airship
+* You can now kick out entities using shift-right click
+* Added fall damage when crashing the aircraft (never lethal by default, configurable)
+* Fixed inventory staying open after entity got removed
+
 # 0.4.2
 
 * Engines now have a fuel buffer and then die slowly when out of fuel